# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: CINTEXA Growth Platform API
  version: 0.2.0
  description: >
    Customer profile, subscriptions, loyalty, contributions, activity, public
    leaderboard, and admin (customers, loyalty adjustment, AI agents)
    endpoints backing the CINTEXA customer portal and admin panel. Customer
    routes authenticate via Clerk session tokens; admin routes authenticate
    via a static x-admin-key header.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/admin/agents/roles:
    get:
      summary: List available AI agent roles (admin)
      operationId: listAgentRoles
      security: [{ adminKey: [] }]
      responses:
        "200":
          description: Array of role identifiers
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles:
                    type: array
                    items: { type: string, enum: [marketing, sales, consumer_simulation, engagement, content_idea] }
        "401":
          description: Missing or invalid admin key
  /api/admin/agents/run:
    post:
      summary: Run one AI agent task synchronously (admin)
      description: >
        Every role is a deterministic, pure function — a foundation ready to
        be swapped for a real LLM-backed implementation later. The
        `engagement` role only produces a scheduling plan; it never executes
        actions against real platforms.
      operationId: runAgentTask
      security: [{ adminKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, input]
              properties:
                role: { type: string, enum: [marketing, sales, consumer_simulation, engagement, content_idea] }
                input:
                  description: Shape depends on `role` — see each role's input schema in artifacts/api-server/src/agents/roles
                  type: object
                requestedByUserId: { type: string }
      responses:
        "200":
          description: The persisted, completed (or failed) task
          content:
            application/json:
              schema:
                type: object
                properties:
                  task: { $ref: "#/components/schemas/AgentTask" }
        "400":
          description: Invalid role or invalid input for that role
        "401":
          description: Missing or invalid admin key
  /api/admin/agents/tasks:
    get:
      summary: List recent agent tasks, optionally filtered by role (admin)
      operationId: listAgentTasks
      security: [{ adminKey: [] }]
      parameters:
        - name: role
          in: query
          schema: { type: string, enum: [marketing, sales, consumer_simulation, engagement, content_idea] }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: A page of task history, most recent first
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items: { $ref: "#/components/schemas/AgentTask" }
                  pagination: { $ref: "#/components/schemas/PaginationMeta" }
        "401":
          description: Missing or invalid admin key
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      description: Clerk session token
    webhookSignature:
      type: apiKey
      in: header
      name: x-webhook-signature
      description: HMAC-SHA256 of the raw request body, signed with WEBHOOK_SECRET
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: Static admin key, matches ADMIN_API_KEY on the server
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
        hasMore: { type: boolean }
    AgentTask:
      type: object
      properties:
        id: { type: string, format: uuid }
        role: { type: string, enum: [marketing, sales, consumer_simulation, engagement, content_idea] }
        status: { type: string, enum: [queued, completed, failed] }
        input: { type: object }
        output: { type: object, nullable: true }
        requestedByUserId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time, nullable: true }